Adjust apache dependency
Prior to this commit, the dependency on puppetlabs-apache was
expressed with a hyphen instead of a forward slash.

This commit corrects the Modulefile and allows the module tool to
consume the mrepo module. Obviously a better fix is to make the
module tool more tolerant but this is a useful short-term fix.